requirements.txt
fastapi
jinja2
mysql-connector-python
pandas
faker
uvicorn
httpx